# generate lpo and vo dependencies for make