--- a/Makefile
+++ b/Makefile
@@ -52,7 +52,8 @@ lpo.mk:
 
 .PHONY: dep-lpo
 dep-lpo:
-	find . -maxdepth 1 -name '*.lp' -exec $(HOL2DK_DIR)/dep-lpo {} \; > lpo.mk
+	cat *.lpo.mk > lpo.mk
+#	find . -maxdepth 1 -name '*.lp' -exec $(HOL2DK_DIR)/dep-lpo {} \; > lpo.mk
 
 .PHONY: clean-dep-lpo
 clean-dep-lpo: