- generate lpo and vo dependencies for make when generating lp files
- xlp.ml and xdk.ml: remove suffix argument from export functions
- use iterators for dependencies
- improve and simplify Makefile: lpo and vo dependencies are automatically computed
- changed prefix of lemmas to "lem" instead of "thm_" before

performance:

hol.ml:
  - without generation: make -j32 lp 47s dep-lpo 42s
  - with generation: make -j32 lp 46s dep-lpo 0.4s

TODO:
- [x] fix .lpo.mk files when calling resize scripts